Build refactor - Create SubjobCalculator class
- Move prepare logic from BuildRequestHandler into Build.
- Move subjob computing logic from BuildRequestHandler into new class.
